Fixed several warnings The biggest difference is changing PresetCollection::m_idx_selected from int to size_t, which fixed many warnings at once.

+++ src/libslic3r/Fill/FillBase.cpp
@@ -611,7 +611,7 @@ static inline SegmentPoint clip_start_segment_and_point(const Points &polyline, 
 	SegmentPoint out;
 	if (polyline.size() >= 2) {
 	    Vec2d pt_prev = polyline.front().cast<double>();
-		for (int i = 1; i < polyline.size(); ++ i) {
+        for (size_t i = 1; i < polyline.size(); ++ i) {
 			Vec2d pt = polyline[i].cast<double>();
 			Vec2d v = pt - pt_prev;
 	        double l2 = v.squaredNorm();
